requirements.txt
google-api-core==1.31.0
google-auth==1.33.0
google-auth-oauthlib==0.4.2
google-cloud-core==1.7.1
google-cloud-pubsub==2.6.1
google-cloud-storage==1.41.0
pandas==1.1.5
requests==2.24.0
protobuf==3.17.3
